feat: restructure API response

+++ src/utils/http_client.py
@@ -3,6 +3,6 @@ import aiohttp
 
 class HttpClientFactory:
     @staticmethod
-    def create(timeout_sec: int) -> aiohttp.ClientSession:
+    def create(timeout_sec: float) -> aiohttp.ClientSession:
         timeout = aiohttp.ClientTimeout(total=timeout_sec)
         return aiohttp.ClientSession(timeout=timeout)
